Don't leave puddles on flat surfaces
If a level 1 liquid block has an adjacent floodable block, we start a
timer for waterfinity_evaporate_interval (default 5 seconds). If after
that time the block is still at level 1 and adjacent to a floodable
block, we set it to an air block.

Instead of tracking that it was the evaporation timer that was fired,
we set a metadata "waterfinity:can_evaporate" which we clear in the
update function, so the block won't evaporate if update touches it.